# LSM: implement move table optimization for compaction