--- a/src/lsm/compaction.zig
+++ b/src/lsm/compaction.zig
@@ -337,6 +337,7 @@ pub fn CompactionType(
                 const level_a = level_b - 1;
 
                 var table_a = compaction.level_a_input.?;
+                assert(table_a.snapshot_max >= snapshot_max);
                 compaction.manifest.move_table(level_a, level_b, snapshot_max, &table_a);
                 assert(table_a.snapshot_max == snapshot_max);