--- a/src/lsm/compaction.zig
+++ b/src/lsm/compaction.zig
@@ -340,13 +340,12 @@ pub fn CompactionType(
                     const level_b = compaction.level_b;
                     const level_a = level_b - 1;
 
-                    var table_a = compaction.level_a_input.?;
+                    const table_a = compaction.level_a_input.?;
                     assert(table_a.snapshot_max >= snapshot_max);
                     assert(compaction.range.table_count == 1);
                     assert(compaction.grid_reservation == null);
 
-                    compaction.manifest.move_table(level_a, level_b, snapshot_max, &table_a);
-                    assert(table_a.snapshot_max == snapshot_max);
+                    compaction.manifest.move_table(level_a, level_b, &table_a);
 
                     compaction.merge_done = true;
                     compaction.status = .done;