When the table can be moved directly between levels without reading values from the iterators, do so and skip the rest of the compaction process.

This is currently a prototype. The benchmark ends up hitting this case quite frequently and not really doing any Grid reads.

## Pre-merge checklist

Performance:

* [x] Compare `zig benchmark` on linux before and after this pr.
    ``` sh
    # benchmark results before
    1000 batches in 252509 ms
    32438 transfers per second
    max p100 latency per 8191 transfers = 1840ms
    
    # benchmark results after
    1000 batches in 109761 ms
    74625 transfers per second
    max p100 latency per 8191 transfers = 563ms
    ```