@jorangreef 2023-02-20
Nice work @kprotty.

What is the batch size used per transfers batch here?