@sentientwaffle 2023-02-24
LGTM. @jorangreef did you want to give it another look before merge?